+++ packages/xgplayer/src/plugin/basePlugin.js
@@ -3,6 +3,7 @@ import Sniffer from '../utils/sniffer'
 import Errors from '../error'
 import * as Events from '../events'
 import XG_DEBUG from '../utils/debug'
+import hooksDescriptor, { hook, useHooks } from '../plugin/hooksDescriptor'
 
 function showErrorMsg (pluginName, msg) {
   console.error(`[${pluginName}] event or callback cant be undefined or null when call ${msg}`)
@@ -37,6 +38,7 @@ class BasePlugin {
     if (Util.checkIsFunction(this.beforeCreate)) {
       this.beforeCreate(args)
     }
+    hooksDescriptor(this)
     /**
      * @private
      */
@@ -208,6 +210,26 @@ class BasePlugin {
    */
   emit (event, res) {
     this.player.emit(event, res)
+  }
+
+  /**
+   * @param { string } hookName
+   * @param { Function } handler
+   * @param { {pre: Function| null , next: Function | null} } preset
+   * @returns
+   */
+  hook (hookName, handler, preset = { pre: null, next: null }) {
+    // eslint-disable-next-line no-return-assign
+    return hook.call(this, ...arguments)
+  }
+
+  /**
+   * @param { string } hookName
+   * @param { Function } handler
+   * @param  {...any} [args]
+   */
+  useHooks (hookName, handler, ...args) {
+    return useHooks.call(this, ...arguments)
   }
 
   /**